Update install.sh
Cambiado a url do respositorio para que sexa https, e comprobar que existen os ficheiros antes de facerlles copia.

+++ install.sh
@@ -3,20 +3,26 @@
 CONFIGDIR=~/.ch01-config
 
 #Download files
-git clone --recurse-submodules git@github.com:chavaone/config-files.git $CONFIGDIR
+git clone --recurse-submodules https://github.com/chavaone/config-files.git $CONFIGDIR
 
 #Install oh-my-zsh
 ln -s $CONFIGDIR/zsh/ch01_theme.zsh-theme $CONFIGDIR/zsh/oh-my-zsh/themes/
 ln -s $CONFIGDIR/zsh/ch01_theme_root.zsh-theme $CONFIGDIR/zsh/oh-my-zsh/themes/
-mv ~/.zshrc ~/.zshrc.old
+if test -f ~/.zshrc; then
+  mv ~/.zshrc ~/.zshrc.old
+fi
 ln -s $CONFIGDIR/zsh/zshrc ~/.zshrc
 
 #Install gitconfig
-mv ~/.gitconfig ~/.gitconfig.old
+if test -f ~/.gitconfig; then
+  mv ~/.gitconfig ~/.gitconfig.old
+fi
 ln -s $CONFIGDIR/git/gitconfig ~/.gitconfig
 
 #Install vimrc
-mv ~/.vimrc ~/.vimrc.old
+if test -f ~/.vimrc; then
+  mv ~/.vimrc ~/.vimrc.old
+fi
 ln -s $CONFIGDIR/vim/vimrc ~/.vimrc
